+++ src/main/java/com/davy/quezzy/controllers/QuestionController.java
@@ -1,8 +1,11 @@
 package com.davy.quezzy.controllers;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import com.davy.quezzy.entities.QuestionEntity;
@@ -22,25 +25,33 @@ public class QuestionController {
 
     @GetMapping
     @Operation(summary = "Listar todas as questões")
-    public List<QuestionEntity> getQuestions() {
-        return questionRepository.findAll();
+    public ResponseEntity<List<QuestionEntity>> getQuestions() {
+        List<QuestionEntity> questions = questionRepository.findAll();
+        return ResponseEntity.ok(questions);
     }
 
     @GetMapping("/{id}")
     @Operation(summary = "Buscar uma questão específica pelo id")
-    public QuestionEntity getQuestionById(@PathVariable Long id) {
-        return questionRepository.findById(id).get();
+    public ResponseEntity<QuestionEntity> getQuestionById(@PathVariable Long id) {    
+        Optional<QuestionEntity> response = questionRepository.findById(id);
+        if(response.isPresent()){
+            QuestionEntity category = response.get();
+            return ResponseEntity.ok(category);
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @GetMapping("/category/{categoryId}")
     @Operation(summary = "Buscar questões por categoria")
-    public List<QuestionEntity> getQuestionsByCategory(@PathVariable Long categoryId) {
-        return questionRepository.findQuestionsByCategoryId(categoryId);
+    public ResponseEntity<List<QuestionEntity>> getQuestionsByCategory(@PathVariable Long categoryId) {
+        List<QuestionEntity> questions = questionRepository.findQuestionsByCategoryId(categoryId);
+        return ResponseEntity.ok(questions);
     }
 
     @PostMapping
     @Operation(summary = "Criar uma nova questão")
-    public QuestionEntity createQuestion(@RequestBody QuestionModel question) {
+    public ResponseEntity<QuestionEntity> createQuestion(@RequestBody QuestionModel question) {
         QuestionEntity questionEntity = new QuestionEntity();
         questionEntity.setQuestion(question.getQuestion());
         questionEntity.setAnswer(question.getAnswer());
@@ -53,29 +64,43 @@ public class QuestionController {
         questionEntity.setWrongAnswer2(question.getWrongAnswer2());
         questionEntity.setWrongAnswer3(question.getWrongAnswer3());
         
-        return questionRepository.save(questionEntity);
+        QuestionEntity createdQuestion = questionRepository.save(questionEntity);
+        return ResponseEntity.status(HttpStatus.CREATED).body(createdQuestion);
     }
 
     @PutMapping("/{id}")
     @Operation(summary = "Atualizar uma questão existente")
-    public QuestionEntity updateQuestion(@PathVariable Long id, @RequestBody QuestionModel question) {
-        QuestionEntity questionToUpdate = this.getQuestionById(id);
-        questionToUpdate.setQuestion(question.getQuestion());
-        questionToUpdate.setAnswer(question.getAnswer());
-        questionToUpdate.setCategoryId(question.getCategoryId());
-        questionToUpdate.setCreatorId(question.getCreatorId());
-        questionToUpdate.setDifficulty(question.getDifficulty());
-        questionToUpdate.setUpdatedAt(DateFormatter.getCurrentDateTime());
-        questionToUpdate.setWrongAnswer1(question.getWrongAnswer1());
-        questionToUpdate.setWrongAnswer2(question.getWrongAnswer2());
-        questionToUpdate.setWrongAnswer3(question.getWrongAnswer3());
-
-        return questionRepository.save(questionToUpdate);
+    public ResponseEntity<QuestionEntity> updateQuestion(@PathVariable Long id, @RequestBody QuestionModel question) {
+        Optional<QuestionEntity> response = questionRepository.findById(id);
+        
+        if(response.isPresent()) {
+            QuestionEntity questionToUpdate = response.get();
+            questionToUpdate.setUpdatedAt(DateFormatter.getCurrentDateTime());
+            questionToUpdate.setQuestion(question.getQuestion());
+            questionToUpdate.setAnswer(question.getAnswer());
+            questionToUpdate.setCategoryId(question.getCategoryId());
+            questionToUpdate.setCreatorId(question.getCreatorId());
+            questionToUpdate.setDifficulty(question.getDifficulty());
+            questionToUpdate.setWrongAnswer1(question.getWrongAnswer1());
+            questionToUpdate.setWrongAnswer2(question.getWrongAnswer2());
+            questionToUpdate.setWrongAnswer3(question.getWrongAnswer3());
+            questionRepository.save(questionToUpdate);
+    
+            return ResponseEntity.ok(questionToUpdate);
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @DeleteMapping("/{id}")
     @Operation(summary = "Deletar uma questão existente")
-    public void deleteQuestion(@PathVariable Long id) {
-        questionRepository.deleteById(id);
+    public ResponseEntity<Void> deleteQuestion(@PathVariable Long id) {
+        Optional<QuestionEntity> response = questionRepository.findById(id);
+        if (response.isPresent()) {
+            questionRepository.deleteById(id);
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 }
