+++ src/main/java/com/davy/quezzy/controllers/UserController.java
@@ -1,8 +1,11 @@
 package com.davy.quezzy.controllers;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import com.davy.quezzy.entities.UserEntity;
@@ -22,44 +25,64 @@ public class UserController {
     
     @GetMapping
     @Operation(summary = "Listar todos os usuários")
-    public List<UserEntity> getUsers() {
-        return userRepository.findAll();
+    public ResponseEntity<List<UserEntity>> getUsers() {
+        List<UserEntity> users = userRepository.findAll();
+        return ResponseEntity.ok(users);
     }
 
     @GetMapping("/{id}")
     @Operation(summary = "Buscar um usuário especifo pelo id")
-    public UserEntity getUserById(@PathVariable Long id) {
-        return userRepository.findById(id).get();
+    public ResponseEntity<UserEntity> getUserById(@PathVariable Long id) {
+        Optional<UserEntity> response = userRepository.findById(id);
+        if(response.isPresent()){
+            UserEntity user = response.get();
+            return ResponseEntity.ok(user);
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @PostMapping
     @Operation(summary = "Criar um novo usuário")
-    public UserEntity createUser(@RequestBody UserModel user) {
+    public ResponseEntity<UserEntity> createUser(@RequestBody UserModel user) {
         UserEntity userEntity = new UserEntity();
+
         userEntity.setUsername(user.getUsername());
         userEntity.setPassword(user.getPassword());
         userEntity.setEmail(user.getEmail());
         userEntity.setCreatedAt(DateFormatter.getCurrentDateTime());
         userEntity.setUpdatedAt(DateFormatter.getCurrentDateTime());
 
-        return userRepository.save(userEntity);
+        UserEntity createdUser = userRepository.save(userEntity);
+        return ResponseEntity.status(HttpStatus.CREATED).body(createdUser);
     }
 
     @PutMapping("/{id}")
     @Operation(summary = "Atualizar um usuário existente")
-    public UserEntity updateUser(@PathVariable Long id, @RequestBody UserModel user) {
-        UserEntity userToUpdate = userRepository.findById(id).get();
-        userToUpdate.setUpdatedAt(DateFormatter.getCurrentDateTime());
-        userToUpdate.setUsername(user.getUsername());
-        userToUpdate.setPassword(user.getPassword());
-        userToUpdate.setEmail(user.getEmail());
-
-        return userRepository.save(userToUpdate);
+    public ResponseEntity<UserEntity> updateUser(@PathVariable Long id, @RequestBody UserModel user) {
+        Optional<UserEntity> response = userRepository.findById(id);
+        if(response.isPresent()) {
+            UserEntity userToUpdate = response.get();
+            userToUpdate.setUsername(user.getUsername());
+            userToUpdate.setPassword(user.getPassword());
+            userToUpdate.setEmail(user.getEmail());
+            userToUpdate.setUpdatedAt(DateFormatter.getCurrentDateTime());
+            UserEntity updatedUser = userRepository.save(userToUpdate);
+            return ResponseEntity.ok(updatedUser);
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @DeleteMapping("/{id}")
     @Operation(summary = "Deletar um usuário existente")
-    public void deleteUser(@PathVariable Long id) {
-        userRepository.deleteById(id);
+    public ResponseEntity<Void> deleteUser(@PathVariable Long id) {
+        Optional<UserEntity> response = userRepository.findById(id);
+        if (response.isPresent()) {
+            userRepository.deleteById(id);
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 }
